% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_activity}
\alias{get_activity}
\title{Read and clean Sunpass activity files}
\usage{
get_activity(
  input_dir = "~/Downloads",
  pattern = "Transactions_transaction_csv_report_"
)
}
\arguments{
\item{input_dir}{the directory in which activity file can be found}

\item{pattern}{the file name pattern for Sunpass Activity files}
}
\value{
a cleaned dataframe of every distinct transaction in every
file in input_dir that matches the pattern
}
\description{
By default, ~/Downloads folder
}
\examples{
\dontrun{
get_activity()
}
}
