% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_cost_by_transponder}
\alias{get_cost_by_transponder}
\title{Get total costs incurred by each transponder}
\usage{
get_cost_by_transponder()
}
\value{
a data from of cost incurred by each transponder
}
\description{
Get total costs incurred by each transponder
}
\examples{
\dontrun{
get_cost_by_transponder()
}
}
