% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_cost_by_trip}
\alias{get_cost_by_trip}
\title{Get cost for each trip in Sunpass data}
\usage{
get_cost_by_trip(gap_between_trips = 5)
}
\arguments{
\item{gap_between_trips}{the minimum number of days required to separate
sequential transactions on a transponder into separate trips}
}
\value{
a data frame of trips with cost, start and end dates
}
\description{
Get cost for each trip in Sunpass data
}
\examples{
\dontrun{
get_cost_by_trip()
get_cost_by_trip(gap_between_trips = 5)
}
}
