% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get a dataframe of all Sunpass data}
\usage{
get_data()
}
\value{
a dataframe of all sunpass data
}
\description{
Read and join activity data to transponder data
}
\examples{
\dontrun{
get_data()
}
}
