% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_with_trip_start_date}
\alias{get_data_with_trip_start_date}
\title{get Sunpass data with an estimated trip_start_date}
\usage{
get_data_with_trip_start_date(gap_between_trips = 5)
}
\arguments{
\item{gap_between_trips}{the minimum number of days required to separate
sequential transactions on a transponder into separate trips}
}
\value{
a data frame of Sunpass transactions with an additional trip_start_date column
}
\description{
Assign a new trip start date for every time a transponder has reported
no activity for \emph{gap_between_trips} days.
}
\examples{
\dontrun{
get_data_with_trip_start_date()
get_data_with_trip_start_date(gap_between_trips = 5)
}
}
