% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_transponder}
\alias{get_transponder}
\title{Read and clean Sunpass transponder files}
\usage{
get_transponder(
  input_dir = "~/Downloads",
  pattern = "my_transponder_csv_report_"
)
}
\arguments{
\item{input_dir}{the directory in which transponder file can be found}

\item{pattern}{the file name pattern for Sunpass transponder files}
}
\value{
a cleaned dataframe of every distinct transponder in every
file in input_dir that matches the pattern
}
\description{
By default, ~/Downloads folder
}
\examples{
\dontrun{
get_transponder()
}
}
